endpoint: Add function to initialize event queue
[ upstream commit 79bf425154292aea71b72eed5e67125f9ed0785e ]

This is useful during endpoint validation when endpoints are being
restored. When they are being restored, their event queue is not yet
initialized because they haven't been exposed to the endpoint manager.
It is important to initialize an endpoint's event queue so that events
are not missed during their restoration.

Signed-off-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>